Add low level orchestrator communication
The Orchestrator is a plain HTTP server that listens
on the default data server port. It is used to
distribute the connection requests to the right kernel.

The connection setup involes the following steps:
* create socket
* parse server hello message
* send http upgrade request

All of this is exposed in the `create_session_client_stream()`
function which returns a socket that is connected to the
requested kernel and can directly be passed to capnp.