Add warning about pycapnp master in README
Currently we are using the master of pycapnp. Unfortunately
pycapnp does not use git for versioning. Therefore, the master
version has the same then the latest released on.
This means it has to be manually installed. To ease the
installation, we have temporarily created a custom index at
https://docs.zhinst.com/pypi/ from which the master version
can be installed.