Adapt the set and get functionality to the latest interface
The Labone capnp session interface lately unified the
behaviour of the setValue and getValue function. The server
can no handle two cases. The first one is single node request
and the second one is a request with a path expression.
In the later case the server resolves the path expression by
its own by calling listNodes internaly.

In the core interface we now expose two set and get functions.
The simple one `get` and `set` take a single path and returns
a single annotated value. The other one accepts a path expression
and returns a list of annotated value. We seperated the two
functionalities in order to have an explicit return type for
both use cases. This hopefully makes the handling of the returned
value more explicit.