Refactor device interface from enum to string
This commit removes the enum for the device interface and replaces it with a plain
string. This is due to the fact the the orchestrator http protocol transports the
interface as a string anyway. In addition new interfaces where introduced and it was
decided that we want to do the parsing of the string within the orchestrator. The
client should therefore no longer care about the user input.

In addition the default was set to an empty string, which will result in the
orchestrator detecting the right interface by its own. This works in most cases
since only one interface is connected most of the times. If more than one
interface is present an internal priority list is applied.