Introduce reflection schema
This commit introduces the reflection schema to the core session.
This reflection schema allow the dynamic loading of all available
capabilities from a server through a single request. The biggest
advantage is that not all used capnp schemas have to be hard coded
in the client.

All the reflection stuff is handled in a separate submodule. The
major change to all other files is to no longer use any hardcoded
schema files but the ones received from the server.