Allow passing a custom queue to the session subscribe
This commit adds the option to add a custom queue class to the
subscribe command on the session. This allows a more flexible
usage of a subscription.

As part of this change a new circular data queue is added which
allows to have a a sort of ring buffer as a queue.