Resolve c++ warning when importing a capnp schema.
This commit is a follow up to the previous commit which tried to catch the
warning. Since the warning is now resolved it no longer needs to be caught.

The kj library (used by capnp) requires the PWD environment variable to be set.
This is needed to resolve relative paths. When using python from the command
line, the PWD variable is set automatically. However, this is not guaranteed.
E.g. when using python from a jupyter notebook, the PWD variable is not set
or set to `/`. This leads to a ugly warning inside the c++ code that can not
be caught using Python's built-in warning handling mechanisms. To avoid this
warning, we temporary set the PWD environment variable to the current working
directory.