Making sure to always use lower paths.
Some devices send paths in lower format, some in upper format, and
sometimes it depends on the method called. This lead to a problem
in finding the right enums for a get request, if the cases didnt
match. To avoid such situations in the future, we should make sure
all paths are in one format (lower case).