Adapt mock to utilize the new schema generation
This commit adapts the mock to utilize the new schema generation.
Pycapnp is not able to provide the node schema as requiered by
the reflection schema. Therefore we have written a small script
to generate the schema in c++ and export it to bytes-packed.
The scripts creates a python module. The module `hpk_schema` is
hardcoded.