Fix circular referencing within the labone.nodetree
This commit adapts the node tree manager to no longer hold
refernces to the created nodes but instead use  weak refernces.

The nodtree manager holds the functionality of creating new nodes
and also the reference to the underlying session. Every node therefore
needs to have a refernce to the manager. To avoid creating n objects for
for the same node the nodes are cached within the nodtree manager. This
caching needs to use weak references otherwise the nodetree will keep
itself alive forever ...