Add a loop manager to manage the capnp event loop lifetime
This commit adds a custom loop manager that manages the lifetime
of the kj eventloop used by capnp. The loop manager exposes a
destroy function that when called prevents any further operation
to capnp. By default this function is registered to the exit of
the asyncio event loop. However if needed the user can create an
instace of the loop manager before creating a session and manage
the lifeime of the loop manager itself. This is especially usefull
when the eventloop is stoped and reentered multiple times.