Upgrade zhinst-comms to 1.1.0
Upgrade to the first "official" version of zhinst-comms. The only
practical change from version 0.0.1 is that 1.1.0 supports string
representation for all types. This is not a breaking change, and
normally a switch of the major version would not be required, but
I wanted to move away from the 0.x versions.