Adapt server to support close and run forever
This commit adapts the capnp server class to support the close
method and run forever method. The close method is used to
stop a running server safely. Since Python does not gurantee that
the destructor is called it is recommended to manage the livetime
by one self. The run forever method is a simple helper function
that is often usefull if a server is running in the main thread.